rework delete-where AST
This commit refactors the deletion structure in the AST to remove
the Delete flag from the pool and instead have a special ast.Delete
object to represent a deletion scan.

This is needed by a subsequent PR that will normalize the "from"
structures to be better aligned wiht SQL syntax.