change the "super query" command to "super"
This commit shortens the "super query" command to "super" so we
can run queries decoupled from the data lake using the top-level
"super" command.  We had to change the charm package a bit so that
the flags defined for super as a standalone command do not conflict
with flags defined by descendent commands and do not interfere with
the help text.