refactor "super dev compile" into "super compile" and "super db compile"
This commit reparents the "super dev compile" command under the top-level
"super" command and creates a parallel command "super db compile".
This makes it much clearer as to when you are compiling a query for
a lake vs a file system.

We also changed the compile commands to take a single query argument
instead of concatenating multiple arguments.  Also, the command options
are a little different: you now say "-dag" if you want the DAG output
and get the AST output otherwise.

There were some inconsitencies in how commands were parented so
we added charm.NoRun() and fixed up the command hierarchy linkages.
The invariant should now be that every internal command properly appears
in the command hierarchy and NoRun() is invoked by the Run() method
of an interior command to give proper help.

The command-line help has also been updated.