add SuperSQL to SuperDB
This commit is a first cut at adding SQL support to SuperDB.

This initial work represents the world's first foundation for
a SQL that blends traditional static analysis of relational queries
(when types are available for the entities queried) with dynamic
execution of polymorphic operators when types are unknown (like SQL++).

It's also the world's first implementation of an (eventually)
SQL-compliant query language embedded within a pipeline-sytle
language that allows aggressive abbrevitions for SQL idioms
to provide lean-forward log-search style experience.  The pipeline
language includes many new operators enabled by super-structured
data that are not available in SQL.

Needless to say, this is also the world's first SQL for
super-structured data.

Pipelines can be added to the end of any SQL expression and
joins are limping along.  SQL unions do not work yet.

While this initial commit provides a very incomplete implementation
of SQL, the foundations are plumbed into the pipeline syntax in a
clean way and should serve as a solid foundation going forward.

It's also all setup for a "binder-in-the-data" approach to
SQL scoping.  More on this coming soon.