unify grammar strings into ast.String
This commit unifies the various types of strings in the PEG grammar to
a single pattern called String and gets rid of ast.QuotedString.  Now that
the locators are all computed from the underlying rule, we can easily
have a single ast.String that represents different forms of strings.
These changes will be used by a subequent PR that will unify the get/file/pool
operators into a single "from", which is in turn needed by the upcoming
SuperSQL additions.