add support for SQL-style case expressions
This commit adds support to the grammar for SQL-style case expressions.
The case-when form is mapped to ast.Conditional and the case-expr-when
form currently reports that it is unsupported in the semantic pass
(we'll need to add runtime support for this other form).

This will be documented when we add docs for SuperSQL.