change zed formats to super formats
This commit changes the formats zson,zng, and vng to jsup, bsup,
and csup, repectively along with all the file extensions in the docs
and tests.  We did not here address -Z, -z, parse_zson, or zjson.

This is primarily a user visible change.  There are still lots of
places in the source code that use the Zed terminology.