simplify file not found messages - less is more
This commit simplifies file-not-found message by dropping the
extra output of the file:/// absolute paths.  We also noticed that
"super:" was prefixed some of the times to errors and not others,
so we took the "super:" prefix out in the one place it was used.

When a data lake object is not found, the full path is printed.
This happens on the vaccum test.  This is not the right UX
when trying to scanned vaccumed parts of the lake, but it's how it
works for now.