rework how command-line file args are wired into DAG
This commit moves the logic that determines if there should be
a null source from queryflags to the compiler.  This decision
really belongs in the semantic pass (as opposed to just after it)
but that change will wait until we evolve the interface to
file storage to handle broader reflection.