allow select operator to receive input from parent
This commit allows the SQL select operator to receive it's
input from it's parent in a pipeline when no from clause is
present.  This means that bare select queries work on constant
values now since the default parent will be a single null value.

We also brought the column naming semantics into closer alignment
with SQL where selected columns also have a string name and not
a dotted field path.

The SQL semantic logic has been simplified and expand further as
we make progress toward our binder-in-the-data approach to
SQL scoping.