simplfy compiler patterns
This commit simplifies the compiler so the code paths are easier to
understand and the API is easier to use.  compiler.Job goes away
as does the variations implementation of the runtime.Compiler.
runtime.Compiler now just has one query method for both local and
lake cases.  We also moved compiler/data.Source to
runtime/exec.Environment as that much more accurately potrays
the role of this struct as representing the execution environment
(lake or local) of the query.  Down the road, we may want to
store config options here (e.g., parallelism, strict compilation,
etc).  We also added the flag -files to "super compile" to ask it
to compile the query assuming there are command-line input files.