add bitwise operators
This commit adds bitwise operators to the PEG grammar, but doesn't
otherwise support them.  If you use them, you'll get an error.

We've been living without bitwise OR in the grammar for a long time
and this time and this change illustates the conflict between "|"
vs "|>" as the pipeline separator.