allow from operators to have parents for robot scans
This commit allows from operators to pull data from its parent
and for each incoming value compute a target to open and read.
This is repeated successively until the parent fails.  When
failures occur, structured errors are propgated downstream.

There is a disabled test mimicing the example in the top-level
README.  It will be enabled in a subsequent upon adding support
for "join using".