disable legacy bitwise-OR syntax in SQL expressions
Ok as disccused, I'm pushing these changes to remove bitwise-OR
from the SQL expression syntax, thus removing the ambiguity with the
SQL shortcuts.  If/when users want to run SuperSQL queries on legacy
SQL with bitwise OR, we can put this grammar back in place but under
a flag to enable it only as needed.  In the meantime, wew ill introduce
bitwise_or(), bitwise_and(), etc named functions in subsequent PR.