bring up JOIN USING and make README query work
This commit gets the using condition for joins working and now allows
the query that is now in the top-level README to work.  We unified
some of the join logic between SQL and SPQ and relaxed some of the
checking of group-by keys with selection, which had been too aggressive.
We also added a test that emulates the README query with smaller data
in local files rather than APIs.