Rename: NullsView to BoolView
Rename NullsView to BoolView since this can be used for creating views
of Bool vectors as well. In addition have all an false value views of
a boolean vector return a non-nil vector. When dealing with boolean values
there's a difference between a false values and a nil vector.

Also have NewView of a boolean vector create a new boolean vector.