Avoid creating empty vectors in vector.Apply
vector.Apply can create a large number of empty (zero-length) vectors
through calls to vector.rip, generating substantial garbage.  Fix that
by changing rip to return nil instead of a slice of empty vectors,
changing Apply to skip the recursive call when it sees a nil from rip
and substitute a nil vector in place of the skipped call's result, and
changing NewTagMap to tolerate nil vectors so NewDynamic will as well.